Extract helpers, combiner ops, panels, notifications, metrics, context menu, and event handler modules (47% reduction to 1,128 lines)
